requirements.txt
flask==2.2.3
python-dotenv==0.19.0
psycopg2-binary==2.9.1
openai==1.0.0
gunicorn==20.1.0
Werkzeug==2.2.3